Code file formatting
Broke the non-arduino specific code into  .c/.h files.  Had to modify some of the arduino code to match the latest versions.

Updated the header template with some copyright information.
